+++ modules/tinymce/src/core/test/ts/browser/fmt/RemoveFormatTest.ts
@@ -1,5 +1,5 @@
 import { context, describe, it } from '@ephox/bedrock-client';
-import { TinyAssertions, TinyHooks, TinySelections } from '@ephox/wrap-mcagar';
+import { TinyAssertions, TinyHooks, TinySelections, TinyState } from '@ephox/wrap-mcagar';
 
 import Editor from 'tinymce/core/api/Editor';
 import { Format } from 'tinymce/core/fmt/FormatTypes';
@@ -259,6 +259,16 @@ describe('browser.tinymce.core.fmt.RemoveFormatTest', () => {
 
       TinyAssertions.assertContent(editor, '<p style="text-align: right;">Test text<img src="link" width="40" height="40"></p>');
       TinyAssertions.assertSelection(editor, [ 0 ], 1, [ 0 ], 2);
+    });
+  });
+
+  it('TINY-9678: Should be a noop if selection is not in an editable context', () => {
+    TinyState.withNoneditableRootEditor(hook.editor(), (editor) => {
+      const initialContent = '<p><strong>test</strong></p><p contenteditable="true"><strong>editable</strong></p>';
+      editor.setContent(initialContent);
+      TinySelections.setSelection(editor, [ 0, 0, 0 ], 4, [ 0, 0, 0 ], 4);
+      editor.formatter.remove('bold');
+      TinyAssertions.assertContent(editor, initialContent);
     });
   });
 });